+++ freemarker-generator-base/src/test/java/org/apache/freemarker/generator/datasource/DatasourceTest.java
@@ -30,12 +30,14 @@ import java.nio.charset.Charset;
 import java.util.Iterator;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
+import static org.apache.freemarker.generator.base.FreeMarkerConstants.DEFAULT_GROUP;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 public class DatasourceTest {
 
+    private static final String ANY_GROUP = "group";
     private static final String ANY_TEXT = "Hello World";
     private static final String ANY_FILE_NAME = "pom.xml";
     private static final Charset ANY_CHAR_SET = UTF_8;
@@ -43,12 +45,14 @@ public class DatasourceTest {
 
     @Test
     public void shouldSupportTextDatasource() throws IOException {
-        try (Datasource datasource = DatasourceFactory.create("stdin", ANY_TEXT)) {
+        try (Datasource datasource = DatasourceFactory.create("stdin", ANY_GROUP, ANY_TEXT)) {
             assertEquals("stdin", datasource.getName());
+            assertEquals(ANY_GROUP, datasource.getGroup());
             assertEquals("stdin", datasource.getBaseName());
             assertEquals("", datasource.getExtension());
             assertEquals("string", datasource.getLocation());
             assertEquals(UTF_8, datasource.getCharset());
+            assertEquals("plain/text", datasource.getContentType());
             assertTrue(datasource.getLength() > 0);
             assertEquals(ANY_TEXT, datasource.getText());
         }
@@ -58,23 +62,27 @@ public class DatasourceTest {
     public void shouldSupportFileDatasource() throws IOException {
         try (Datasource datasource = DatasourceFactory.create(ANY_FILE, ANY_CHAR_SET)) {
             assertEquals(ANY_FILE_NAME, datasource.getName());
+            assertEquals(DEFAULT_GROUP, datasource.getGroup());
             assertEquals("pom", datasource.getBaseName());
             assertEquals("xml", datasource.getExtension());
             assertEquals(ANY_FILE.getAbsolutePath(), datasource.getLocation());
             assertEquals(Charset.defaultCharset(), datasource.getCharset());
+            assertEquals("application/xml", datasource.getContentType());
             assertTrue(datasource.getLength() > 0);
             assertFalse(datasource.getText().isEmpty());
         }
     }
 
-    @Ignore
+    @Ignore("Requires internet conenection")
     @Test
     public void shouldSupportUrlDatasource() throws IOException {
         try (Datasource datasource = DatasourceFactory.create(new URL("https://google.com?foo=bar"))) {
             assertEquals("google.com", datasource.getName());
+            assertEquals(DEFAULT_GROUP, datasource.getGroup());
             assertEquals("google", datasource.getBaseName());
             assertEquals("com", datasource.getExtension());
             assertEquals("https://google.com", datasource.getLocation());
+            assertEquals("text/html; charset=ISO-8859-1", datasource.getContentType());
             assertEquals(UTF_8, datasource.getCharset());
             assertEquals(-1, datasource.getLength());
             assertFalse(datasource.getText().isEmpty());
@@ -127,7 +135,7 @@ public class DatasourceTest {
     }
 
     private static Datasource textDatasource() {
-        return DatasourceFactory.create("stdin", ANY_TEXT);
+        return DatasourceFactory.create("stdin", "default", ANY_TEXT);
     }
 
     private static final class TestClosable implements Closeable {